pyproject.toml
[tool.poetry]
name = "kleiderkammer"
version = "0.1.0"
description = ""
authors = ["Michel Blank"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
uWSGI = "^2.0.20"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.6.2"
Werkzeug = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
